Create __init__.py
Hey there, Bats3c, love the dope tool. This was exactly what I needed for a pentest I was just working on. Ran into a tiny problem with an easy fix and thought I'd point it out.

There should be a __init__.py file in the lib directory so that it doesn't cause sys.path and import conflicts. For example, if you have another tool that uses `lib/` as a directory in its structure, and you install it into your `python3.x/site-packages` directory, it will prefer to import lib from there instead of the local directory like you want. It can result in some confusing errors for those not versed in the ways of python. 

Feel free to close this and patch it in yourself. Just an empty __init__.py file was enough to solve the issue for me :)